APPDEV-1166 code refactoring (#533)
* APPDEV-1166 Package frinends to renamed to friends.

* APPDEV-1166 Notification messages pagination error fixes.
-- Removed the paging related query parameters for notification message API calls and relying on Next and Previous links.

* APPDEV-1166 Refactoring of NotificationManagerImpl class.
