+++ app/docs/nu/yona/app/ui/friends/OverviewFragment.html
@@ -3,7 +3,7 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (1.8.0_25) on Sun May 01 13:07:48 IST 2016 -->
-<title>FriendsFragment</title>
+<title>OverviewFragment</title>
 <meta name="date" content="2016-05-01">
 <link rel="stylesheet" type="text/css" href="../../../../../stylesheet.css" title="Style">
 <script type="text/javascript" src="../../../../../script.js"></script>
@@ -12,13 +12,13 @@
 <script type="text/javascript"><!--
     try {
         if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="FriendsFragment";
+            parent.document.title="OverviewFragment";
         }
     }
     catch(err) {
     }
 //-->
-var methods = {"i0":10,"i1":10};
+var methods = {"i0":10};
 var tabs = {65535:["t0","All Methods"],2:["t2","Instance Methods"],8:["t4","Concrete Methods"]};
 var altColor = "altColor";
 var rowColor = "rowColor";
@@ -48,12 +48,12 @@ var activeTableTab = "activeTableTab";
 </div>
 <div class="subNav">
 <ul class="navList">
-<li><a href="../../../../../nu/yona/app/ui/frinends/AddFriendManually.html" title="class in nu.yona.app.ui.frinends"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
-<li><a href="../../../../../nu/yona/app/ui/frinends/OverviewFragment.html" title="class in nu.yona.app.ui.frinends"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
+<li><a href="../../../../../nu/yona/app/ui/frinends/FriendsFragment.html" title="class in nu.yona.app.ui.friends"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
+<li><a href="../../../../../nu/yona/app/ui/frinends/TimelineFragment.html" title="class in nu.yona.app.ui.friends"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
 </ul>
 <ul class="navList">
-<li><a href="../../../../../index.html?nu/yona/app/ui/frinends/FriendsFragment.html" target="_top">Frames</a></li>
-<li><a href="FriendsFragment.html" target="_top">No&nbsp;Frames</a></li>
+<li><a href="../../../../../index.html?nu/yona/app/ui/frinends/OverviewFragment.html" target="_top">Frames</a></li>
+<li><a href="OverviewFragment.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_top">
 <li><a href="../../../../../allclasses-noframe.html">All&nbsp;Classes</a></li>
@@ -91,8 +91,8 @@ var activeTableTab = "activeTableTab";
 <!-- ========= END OF TOP NAVBAR ========= -->
 <!-- ======== START OF CLASS DATA ======== -->
 <div class="header">
-<div class="subTitle">nu.yona.app.ui.frinends</div>
-<h2 title="Class FriendsFragment" class="title">Class FriendsFragment</h2>
+<div class="subTitle">nu.yona.app.ui.friends</div>
+<h2 title="Class OverviewFragment" class="title">Class OverviewFragment</h2>
 </div>
 <div class="contentContainer">
 <ul class="inheritance">
@@ -102,7 +102,7 @@ var activeTableTab = "activeTableTab";
 <li><a href="../../../../../nu/yona/app/ui/BaseFragment.html" title="class in nu.yona.app.ui">nu.yona.app.ui.BaseFragment</a></li>
 <li>
 <ul class="inheritance">
-<li>nu.yona.app.ui.frinends.FriendsFragment</li>
+<li>nu.yona.app.ui.friends.OverviewFragment</li>
 </ul>
 </li>
 </ul>
@@ -113,7 +113,7 @@ var activeTableTab = "activeTableTab";
 <li class="blockList">
 <hr>
 <br>
-<pre>public class <span class="typeNameLabel">FriendsFragment</span>
+<pre>public class <span class="typeNameLabel">OverviewFragment</span>
 extends <a href="../../../../../nu/yona/app/ui/BaseFragment.html" title="class in nu.yona.app.ui">BaseFragment</a></pre>
 <div class="block">Created by kinnarvasa on 21/03/16.</div>
 </li>
@@ -134,7 +134,7 @@ extends <a href="../../../../../nu/yona/app/ui/BaseFragment.html" title="class i
 <th class="colOne" scope="col">Constructor and Description</th>
 </tr>
 <tr class="altColor">
-<td class="colOne"><code><span class="memberNameLink"><a href="../../../../../nu/yona/app/ui/frinends/FriendsFragment.html#FriendsFragment--">FriendsFragment</a></span>()</code>&nbsp;</td>
+<td class="colOne"><code><span class="memberNameLink"><a href="../../../../../nu/yona/app/ui/frinends/OverviewFragment.html#OverviewFragment--">OverviewFragment</a></span>()</code>&nbsp;</td>
 </tr>
 </table>
 </li>
@@ -153,13 +153,9 @@ extends <a href="../../../../../nu/yona/app/ui/BaseFragment.html" title="class i
 </tr>
 <tr id="i0" class="altColor">
 <td class="colFirst"><code>View</code></td>
-<td class="colLast"><code><span class="memberNameLink"><a href="../../../../../nu/yona/app/ui/frinends/FriendsFragment.html#onCreateView-LayoutInflater-ViewGroup-Bundle-">onCreateView</a></span>(LayoutInflater&nbsp;inflater,
+<td class="colLast"><code><span class="memberNameLink"><a href="../../../../../nu/yona/app/ui/frinends/OverviewFragment.html#onCreateView-LayoutInflater-ViewGroup-Bundle-">onCreateView</a></span>(LayoutInflater&nbsp;inflater,
             ViewGroup&nbsp;container,
             Bundle&nbsp;savedInstanceState)</code>&nbsp;</td>
-</tr>
-<tr id="i1" class="rowColor">
-<td class="colFirst"><code>void</code></td>
-<td class="colLast"><code><span class="memberNameLink"><a href="../../../../../nu/yona/app/ui/frinends/FriendsFragment.html#onResume--">onResume</a></span>()</code>&nbsp;</td>
 </tr>
 </table>
 <ul class="blockList">
@@ -183,13 +179,13 @@ extends <a href="../../../../../nu/yona/app/ui/BaseFragment.html" title="class i
 <!--   -->
 </a>
 <h3>Constructor Detail</h3>
-<a name="FriendsFragment--">
+<a name="OverviewFragment--">
 <!--   -->
 </a>
 <ul class="blockListLast">
 <li class="blockList">
-<h4>FriendsFragment</h4>
-<pre>public&nbsp;FriendsFragment()</pre>
+<h4>OverviewFragment</h4>
+<pre>public&nbsp;OverviewFragment()</pre>
 </li>
 </ul>
 </li>
@@ -203,7 +199,7 @@ extends <a href="../../../../../nu/yona/app/ui/BaseFragment.html" title="class i
 <a name="onCreateView-LayoutInflater-ViewGroup-Bundle-">
 <!--   -->
 </a>
-<ul class="blockList">
+<ul class="blockListLast">
 <li class="blockList">
 <h4>onCreateView</h4>
 <pre>@Nullable
@@ -212,15 +208,6 @@ public&nbsp;View&nbsp;onCreateView(LayoutInflater&nbsp;inflater,
                                    ViewGroup&nbsp;container,
                                    @Nullable
                                    Bundle&nbsp;savedInstanceState)</pre>
-</li>
-</ul>
-<a name="onResume--">
-<!--   -->
-</a>
-<ul class="blockListLast">
-<li class="blockList">
-<h4>onResume</h4>
-<pre>public&nbsp;void&nbsp;onResume()</pre>
 </li>
 </ul>
 </li>
@@ -250,12 +237,12 @@ public&nbsp;View&nbsp;onCreateView(LayoutInflater&nbsp;inflater,
 </div>
 <div class="subNav">
 <ul class="navList">
-<li><a href="../../../../../nu/yona/app/ui/frinends/AddFriendManually.html" title="class in nu.yona.app.ui.frinends"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
-<li><a href="../../../../../nu/yona/app/ui/frinends/OverviewFragment.html" title="class in nu.yona.app.ui.frinends"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
+<li><a href="../../../../../nu/yona/app/ui/frinends/FriendsFragment.html" title="class in nu.yona.app.ui.friends"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
+<li><a href="../../../../../nu/yona/app/ui/frinends/TimelineFragment.html" title="class in nu.yona.app.ui.friends"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
 </ul>
 <ul class="navList">
-<li><a href="../../../../../index.html?nu/yona/app/ui/frinends/FriendsFragment.html" target="_top">Frames</a></li>
-<li><a href="FriendsFragment.html" target="_top">No&nbsp;Frames</a></li>
+<li><a href="../../../../../index.html?nu/yona/app/ui/frinends/OverviewFragment.html" target="_top">Frames</a></li>
+<li><a href="OverviewFragment.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_bottom">
 <li><a href="../../../../../allclasses-noframe.html">All&nbsp;Classes</a></li>
