+++ app/docs/nu/yona/app/ui/friends/package-tree.html
@@ -3,7 +3,7 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (1.8.0_25) on Sun May 01 13:07:49 IST 2016 -->
-<title>nu.yona.app.ui.frinends</title>
+<title>nu.yona.app.ui.friends Class Hierarchy</title>
 <meta name="date" content="2016-05-01">
 <link rel="stylesheet" type="text/css" href="../../../../../stylesheet.css" title="Style">
 <script type="text/javascript" src="../../../../../script.js"></script>
@@ -12,13 +12,12 @@
 <script type="text/javascript"><!--
     try {
         if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="nu.yona.app.ui.frinends";
+            parent.document.title="nu.yona.app.ui.friends Class Hierarchy";
         }
     }
     catch(err) {
     }
-//-->
-</script>
+//--></script>
 <noscript>
 <div>JavaScript is disabled on your browser.</div>
 </noscript>
@@ -32,9 +31,9 @@
 </a>
 <ul class="navList" title="Navigation">
 <li><a href="../../../../../overview-summary.html">Overview</a></li>
-<li class="navBarCell1Rev">Package</li>
+<li><a href="package-summary.html">Package</a></li>
 <li>Class</li>
-<li><a href="package-tree.html">Tree</a></li>
+<li class="navBarCell1Rev">Tree</li>
 <li><a href="../../../../../deprecated-list.html">Deprecated</a></li>
 <li><a href="../../../../../index-files/index-1.html">Index</a></li>
 <li><a href="../../../../../help-doc.html">Help</a></li>
@@ -42,12 +41,12 @@
 </div>
 <div class="subNav">
 <ul class="navList">
-<li><a href="../../../../../nu/yona/app/ui/dashboard/package-summary.html">Prev&nbsp;Package</a></li>
-<li><a href="../../../../../nu/yona/app/ui/login/package-summary.html">Next&nbsp;Package</a></li>
+<li><a href="../../../../../nu/yona/app/ui/dashboard/package-tree.html">Prev</a></li>
+<li><a href="../../../../../nu/yona/app/ui/login/package-tree.html">Next</a></li>
 </ul>
 <ul class="navList">
-<li><a href="../../../../../index.html?nu/yona/app/ui/frinends/package-summary.html" target="_top">Frames</a></li>
-<li><a href="package-summary.html" target="_top">No&nbsp;Frames</a></li>
+<li><a href="../../../../../index.html?nu/yona/app/ui/frinends/package-tree.html" target="_top">Frames</a></li>
+<li><a href="package-tree.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_top">
 <li><a href="../../../../../allclasses-noframe.html">All&nbsp;Classes</a></li>
@@ -69,50 +68,27 @@
 </a></div>
 <!-- ========= END OF TOP NAVBAR ========= -->
 <div class="header">
-<h1 title="Package" class="title">Package&nbsp;nu.yona.app.ui.frinends</h1>
+<h1 class="title">Hierarchy For Package nu.yona.app.ui.friends</h1>
+<span class="packageHierarchyLabel">Package Hierarchies:</span>
+<ul class="horizontal">
+<li><a href="../../../../../overview-tree.html">All Packages</a></li>
+</ul>
 </div>
 <div class="contentContainer">
-<ul class="blockList">
-<li class="blockList">
-<table class="typeSummary" border="0" cellpadding="3" cellspacing="0" summary="Class Summary table, listing classes, and an explanation">
-<caption><span>Class Summary</span><span class="tabEnd">&nbsp;</span></caption>
-<tr>
-<th class="colFirst" scope="col">Class</th>
-<th class="colLast" scope="col">Description</th>
-</tr>
-<tbody>
-<tr class="altColor">
-<td class="colFirst"><a href="../../../../../nu/yona/app/ui/frinends/AddFriendFragment.html" title="class in nu.yona.app.ui.frinends">AddFriendFragment</a></td>
-<td class="colLast">
-<div class="block">Created by kinnarvasa on 27/04/16.</div>
-</td>
-</tr>
-<tr class="rowColor">
-<td class="colFirst"><a href="../../../../../nu/yona/app/ui/frinends/AddFriendManually.html" title="class in nu.yona.app.ui.frinends">AddFriendManually</a></td>
-<td class="colLast">
-<div class="block">Created by kinnarvasa on 27/04/16.</div>
-</td>
-</tr>
-<tr class="altColor">
-<td class="colFirst"><a href="../../../../../nu/yona/app/ui/frinends/FriendsFragment.html" title="class in nu.yona.app.ui.frinends">FriendsFragment</a></td>
-<td class="colLast">
-<div class="block">Created by kinnarvasa on 21/03/16.</div>
-</td>
-</tr>
-<tr class="rowColor">
-<td class="colFirst"><a href="../../../../../nu/yona/app/ui/frinends/OverviewFragment.html" title="class in nu.yona.app.ui.frinends">OverviewFragment</a></td>
-<td class="colLast">
-<div class="block">Created by kinnarvasa on 21/03/16.</div>
-</td>
-</tr>
-<tr class="altColor">
-<td class="colFirst"><a href="../../../../../nu/yona/app/ui/frinends/TimelineFragment.html" title="class in nu.yona.app.ui.frinends">TimelineFragment</a></td>
-<td class="colLast">
-<div class="block">Created by kinnarvasa on 21/03/16.</div>
-</td>
-</tr>
-</tbody>
-</table>
+<h2 title="Class Hierarchy">Class Hierarchy</h2>
+<ul>
+<li type="circle">java.lang.Object
+<ul>
+<li type="circle">nu.yona.app.ui.<a href="../../../../../nu/yona/app/ui/BaseFragment.html" title="class in nu.yona.app.ui"><span class="typeNameLink">BaseFragment</span></a>
+<ul>
+<li type="circle">nu.yona.app.ui.frinends.<a href="../../../../../nu/yona/app/ui/frinends/AddFriendFragment.html" title="class in nu.yona.app.ui.friends"><span class="typeNameLink">AddFriendFragment</span></a></li>
+<li type="circle">nu.yona.app.ui.frinends.<a href="../../../../../nu/yona/app/ui/frinends/AddFriendManually.html" title="class in nu.yona.app.ui.friends"><span class="typeNameLink">AddFriendManually</span></a></li>
+<li type="circle">nu.yona.app.ui.frinends.<a href="../../../../../nu/yona/app/ui/frinends/FriendsFragment.html" title="class in nu.yona.app.ui.friends"><span class="typeNameLink">FriendsFragment</span></a></li>
+<li type="circle">nu.yona.app.ui.frinends.<a href="../../../../../nu/yona/app/ui/frinends/OverviewFragment.html" title="class in nu.yona.app.ui.friends"><span class="typeNameLink">OverviewFragment</span></a></li>
+<li type="circle">nu.yona.app.ui.frinends.<a href="../../../../../nu/yona/app/ui/frinends/TimelineFragment.html" title="class in nu.yona.app.ui.friends"><span class="typeNameLink">TimelineFragment</span></a></li>
+</ul>
+</li>
+</ul>
 </li>
 </ul>
 </div>
@@ -126,9 +102,9 @@
 </a>
 <ul class="navList" title="Navigation">
 <li><a href="../../../../../overview-summary.html">Overview</a></li>
-<li class="navBarCell1Rev">Package</li>
+<li><a href="package-summary.html">Package</a></li>
 <li>Class</li>
-<li><a href="package-tree.html">Tree</a></li>
+<li class="navBarCell1Rev">Tree</li>
 <li><a href="../../../../../deprecated-list.html">Deprecated</a></li>
 <li><a href="../../../../../index-files/index-1.html">Index</a></li>
 <li><a href="../../../../../help-doc.html">Help</a></li>
@@ -136,12 +112,12 @@
 </div>
 <div class="subNav">
 <ul class="navList">
-<li><a href="../../../../../nu/yona/app/ui/dashboard/package-summary.html">Prev&nbsp;Package</a></li>
-<li><a href="../../../../../nu/yona/app/ui/login/package-summary.html">Next&nbsp;Package</a></li>
+<li><a href="../../../../../nu/yona/app/ui/dashboard/package-tree.html">Prev</a></li>
+<li><a href="../../../../../nu/yona/app/ui/login/package-tree.html">Next</a></li>
 </ul>
 <ul class="navList">
-<li><a href="../../../../../index.html?nu/yona/app/ui/frinends/package-summary.html" target="_top">Frames</a></li>
-<li><a href="package-summary.html" target="_top">No&nbsp;Frames</a></li>
+<li><a href="../../../../../index.html?nu/yona/app/ui/frinends/package-tree.html" target="_top">Frames</a></li>
+<li><a href="package-tree.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_bottom">
 <li><a href="../../../../../allclasses-noframe.html">All&nbsp;Classes</a></li>
