+++ app/src/main/java/nu/yona/app/ui/friends/FriendsRequestFragment.java
@@ -6,7 +6,7 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-package nu.yona.app.ui.frinends;
+package nu.yona.app.ui.friends;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
@@ -214,7 +214,7 @@ public class FriendsRequestFragment extends BaseProfileFragment implements View.
 		MessageBody messageBody = new MessageBody();
 		messageBody.setProperties(new Properties());
 		YonaActivity.getActivity().showLoadingView(true, null);
-		APIManager.getInstance().getNotificationManager().postMessage(url, messageBody, 0, 0, new DataLoadListener()
+		APIManager.getInstance().getNotificationManager().postMessage(url, messageBody, new DataLoadListener()
 		{
 			@Override
 			public void onDataLoad(Object result)
